+++ lib/features/auth/controller/auth_controller.dart
@@ -1,15 +1,16 @@
 // auth_controller.dart
 import 'package:devotion/core/util/utils.dart';
 import 'package:devotion/features/auth/data/models/user_models.dart';
-import 'package:devotion/features/auth/data/repository/auth_repository.dart';
+import 'package:devotion/features/auth/Repository/auth_repository.dart';
 import 'package:devotion/features/auth/presentation/screen/welcome.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter/material.dart';
 
+//use the userProvider to update user information
 final userProvider = StateProvider<UserModel?>((ref) => null);
 
-final authControllerProvider = StateNotifierProvider<AuthController, bool>(
+final authControllerProvider = StateNotifierProvider<AuthController, bool>( 
   (ref) => AuthController(
     authRepository: ref.watch(authRepositoryProvider),
     ref: ref,
@@ -21,14 +22,12 @@ final authStateChangeProvider = StreamProvider((ref) {
   return authController.authStateChange;
 });
 
-
-
 final getUserDataProvider = StreamProvider.family((ref, String uid) {
   final authController = ref.watch(authControllerProvider.notifier);
   return authController.getUserData(uid);
 });
 
- class AuthController extends StateNotifier<bool> {
+class AuthController extends StateNotifier<bool> {
   final AuthRepository _authRepository;
   final Ref _ref;
 
@@ -55,44 +54,13 @@ final getUserDataProvider = StreamProvider.family((ref, String uid) {
       MaterialPageRoute(builder: (context) => const WelcomeScreen()),
     );
   }
+ 
+
+
+
 
   Stream<UserModel> getUserData(String uid) {
     return _authRepository.getUserData(uid);
-  }
-
-  // Sign up with phone number
-  void signUpWithPhoneNumber(String phoneNumber, BuildContext context) {
-    _authRepository.signUpWithPhoneNumber(phoneNumber, context);
-  }
-
-  // Verify 6-digit code for phone authentication
-  void verifyCode(String verificationId, String code, BuildContext context) {
-    _authRepository.verifyCode(verificationId, code, context);
-  }
-
-  // Sign up with userEmail and password
-  void signUpWithEmail(
-      String userEmail, String password, String role, BuildContext context) {
-    _authRepository.signUpWithEmail(userEmail, password, role, context);
-  }
-
-  // Sign in with userEmail and password
-  void signInWithEmail(
-      String userEmail, String password, BuildContext context) async {
-    final user =
-        await _authRepository.signInWithEmail(userEmail, password, context);
-
-    if (user != null) {
-      Navigator.pushReplacement(
-        context,
-        MaterialPageRoute(builder: (context) => const WelcomeScreen()),
-      );
-    }
-  }
-
-  // Send password reset userEmail
-  void sendPasswordReset(String userEmail, BuildContext context) {
-    _authRepository.sendPasswordReset(userEmail, context);
   }
 
   // Sign out the user
@@ -101,4 +69,43 @@ final getUserDataProvider = StreamProvider.family((ref, String uid) {
     Navigator.pushReplacementNamed(
         context, '/login'); // Update to navigate back to the login screen
   }
+
+ 
+
+  // // Sign up with phone number
+  // void signUpWithPhoneNumber(String phoneNumber, BuildContext context) {
+  //   _authRepository.signUpWithPhoneNumber(phoneNumber, context);
+  // }
+
+  // // Verify 6-digit code for phone authentication
+  // void verifyCode(String verificationId, String code, BuildContext context) {
+  //   _authRepository.verifyCode(verificationId, code, context);
+  // }
+
+  // // Sign up with userEmail and password
+  // void signUpWithEmail(
+  //     String userEmail, String password, String role, BuildContext context) {
+  //   _authRepository.signUpWithEmail(userEmail, password, role, context);
+  // }
+
+  // // Sign in with userEmail and password
+  // void signInWithEmail(
+  //     String userEmail, String password, BuildContext context) async {
+  //   final user =
+  //       await _authRepository.signInWithEmail(userEmail, password, context);
+
+  //   if (user != null) {
+  //     Navigator.pushReplacement(
+  //       context,
+  //       MaterialPageRoute(builder: (context) => const WelcomeScreen()),
+  //     );
+  //   }
+  // }
+
+  // // Send password reset userEmail
+  // void sendPasswordReset(String userEmail, BuildContext context) {
+  //   _authRepository.sendPasswordReset(userEmail, context);
+  // }
+
+  
 }
